+++ packages/workeros-programs/src/node/require-runtime.js
@@ -14,6 +14,7 @@ import { createUrl } from "./url.js";
 import { createFs } from "./fs.js";
 import { createModule } from "./module.js";
 import { buffer as bufferModule } from "./buffer.js";
+import { assert as assertModule, strict as assertStrictModule } from "./assert.js";
 import { stringDecoder as stringDecoderModule } from "./string_decoder.js";
 import { EventEmitter } from "./events.js";
 import { util as utilModule } from "./util.js";
@@ -56,6 +57,8 @@ export function makeBuiltins(sys, extras) {
     ["os", os],
     ["url", url],
     ["buffer", bufferModule],
+    ["assert", assertModule],
+    ["assert/strict", assertStrictModule],
     ["string_decoder", stringDecoderModule],
     ["events", EventEmitter],
     ["util", utilModule],